requirements.txt
customtkinter
psutil
Pillow
opencv-python
pywin32